+++ openslides/motions/static/js/motions/site.js
@@ -442,8 +442,9 @@ angular.module('OpenSlidesApp.motions.site', [
     'Workflow',
     'Agenda',
     'AgendaTree',
-    function ($filter, gettextCatalog, operator, Editor, MotionComment, Category,
-        Config, Mediafile, MotionBlock, Tag, User, Workflow, Agenda, AgendaTree) {
+    'ShowAsAgendaItemField',
+    function ($filter, gettextCatalog, operator, Editor, MotionComment, Category, Config,
+        Mediafile, MotionBlock, Tag, User, Workflow, Agenda, AgendaTree, ShowAsAgendaItemField) {
         return {
             // ngDialog for motion form
             // If motion is given and not null, we're editing an already existing motion
@@ -467,6 +468,10 @@ angular.module('OpenSlidesApp.motions.site', [
             },
             // angular-formly fields for motion form
             getFormFields: function (isCreateForm, isParagraphBasedAmendment) {
+                if (!isParagraphBasedAmendment) { // catch null and undefined. Angular formy doesn't like this.
+                    isParagraphBasedAmendment = false;
+                }
+
                 var workflows = Workflow.getAll();
                 var images = Mediafile.getAllImages();
                 var formFields = [];
@@ -540,15 +545,7 @@ angular.module('OpenSlidesApp.motions.site', [
 
                 // show as agenda item + parent item
                 if (isCreateForm) {
-                    formFields.push({
-                        key: 'showAsAgendaItem',
-                        type: 'checkbox',
-                        templateOptions: {
-                            label: gettextCatalog.getString('Show as agenda item'),
-                            description: gettextCatalog.getString('If deactivated the motion appears as internal item on agenda.')
-                        },
-                        hide: !(operator.hasPerms('motions.can_manage') && operator.hasPerms('agenda.can_manage'))
-                    });
+                    formFields.push(ShowAsAgendaItemField('motions.can_manage'));
                     formFields.push({
                         key: 'agenda_parent_id',
                         type: 'select-single',
@@ -1215,7 +1212,14 @@ angular.module('OpenSlidesApp.motions.site', [
             return Motion.lastModified();
         }, function () {
             // get all main motions and order by identifier (after custom ordering)
-            $scope.motions = _.orderBy(Motion.filter({parent_id: undefined}), ['identifier']);
+            var motions;
+            if (Config.get('motions_amendments_main_table').value) {
+                motions = Motion.getAll();
+            } else {
+                motions = Motion.filter({parent_id: undefined});
+            }
+
+            $scope.motions = _.orderBy(motions, ['identifier']);
             _.forEach($scope.motions, function (motion) {
                 MotionComment.populateFields(motion);
                 motion.personalNote = PersonalNoteManager.getNote(motion);
@@ -2223,7 +2227,10 @@ angular.module('OpenSlidesApp.motions.site', [
         User.bindAll({}, $scope, 'users');
         Workflow.bindAll({}, $scope, 'workflows');
 
-        $scope.model = {};
+        $scope.model = {
+            agenda_type: parseInt(Config.get('agenda_new_items_default_visibility').value),
+        };
+
         $scope.alert = {};
 
         // Check whether this is a new amendment.
@@ -2279,8 +2286,6 @@ angular.module('OpenSlidesApp.motions.site', [
 
         // save motion
         $scope.save = function (motion, gotoDetailView) {
-            motion.agenda_type = motion.showAsAgendaItem ? 1 : 2;
-
             if (isAmendment && motion.paragraphNo !== undefined) {
                 var orig_paragraphs = parentMotion.getTextParagraphs(parentMotion.active_version, false);
                 motion.amendment_paragraphs = orig_paragraphs.map(function (_, idx) {
@@ -3235,6 +3240,7 @@ angular.module('OpenSlidesApp.motions.site', [
         // subgroup Amendments
         gettext('Amendments');
         gettext('Activate amendments');
+        gettext('Show amendments together with motions');
         gettext('Prefix for the identifier for amendments');
         gettext('Apply text for new amendments');
         gettext('The title of the motion is always applied.');